Use unsigned long for pointer

+++ sys/kprintf.c
@@ -53,7 +53,7 @@ void print_character(char c) {
     }
 }
 
-char *decimal_conversion(unsigned int decimal, int base) {
+char *decimal_conversion(unsigned long decimal, int base) {
     static char buf[SIZE];
     char *result = &buf[SIZE - 1], representation[] = "0123456789ABCDEF";
     int remainder;
@@ -91,7 +91,7 @@ void output(const char *fmt, va_list arguments) {
                     break;
 
                 case 'x':
-                    kprintf(decimal_conversion(va_arg(arguments, unsigned int), 16));
+                    kprintf(decimal_conversion(va_arg(arguments, unsigned long), 16));
                     break;
 
                 case 's':
@@ -100,7 +100,7 @@ void output(const char *fmt, va_list arguments) {
 
                 case 'p':
                     kprintf("0x");
-                    kprintf(decimal_conversion(va_arg(arguments, unsigned int), 16));
+                    kprintf(decimal_conversion(va_arg(arguments, unsigned long), 16));
                     break;
 
                 default:
